requirements.txt
requests
openai
dotenv
schedule
snowflake-connector-python